Added draw lives, enemies to screen
Used some stuff from demo02-3 so that lives and enemies are drawn to new
rooms.  Also, I think I figured out why it is only working for the top
half: for bottom, it uses collision_bot, not collision.  I added the
code from collision to collision_bot, but collision_bot is pretty much a
duplicate of collision so let's optimize this to save space.  HOWEVER,
it seems to be working for doors!