Storing floors in 8 bytes
Working on storing floors in 8 bytes.  Broke the game, which is why it's
in a different file. Working on code from 158-231.  Not efficient nor
fully functional, but does compile and prints out the basic idea.  If we
store floors in 8 bytes we sacrifice some detail but save a lot of
memory (we can store detail in the 8th byte for floors, see code).

But yeah, definitely need help on figuring this out!