Random Portal Generation and Doors
Doors are being stupid, there is an error somewhere and I can't find it.
Portal generation works, just random.  We can set portals if we want,
just need to store in a byte somewhere.  Gave up on trying to make
levels repeat/redraw rows to save bytes because of error that is
probably the one that is occuring with doors right now.