Added title page, updated random function
Made random subroutine work with clock; x and y variables remain
unchanged upon completion of the subroutine, so nothing needs to be
changed in that regard.

Added title page (just blank screen, no names at the moment: f3 to start
game) in order to create the "randomness" for the random generator.
Thus, random enemy spawning at the beginning of the game.

Also, enemies should be different colours now.   Not the prettiest
different colours, but different colours.